+++ src/components/style/CommunityCardStyle.tsx
