+++ src/pages/_app.tsx
@@ -2,8 +2,8 @@ import '@/styles/globals.css';
 
 import { AppProps } from 'next/app';
 
-import Footer from '@/components/Footer';
-import Header from '@/components/Header';
+import Footer from '@/components/layout/Footer';
+import Header from '@/components/layout/Header';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
